LayoutTests/css3/flexbox/css-properties-expected.txt

Tests being able to set the display to -webkit-flexbox and -webkit-inline-flexbox.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.style.display is "-webkit-flexbox"
PASS div.style.display is "inline"
PASS div.style.display is "-webkit-inline-flexbox"
PASS div.style.display is "-webkit-inline-flexbox"
PASS div.style.display is "block"
PASS flexitem.style.webkitFlexOrder is ""
PASS flexitem.style.webkitFlexOrder is "2"
PASS flexitem.style.webkitFlexOrder is "-1"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is ""
PASS successfullyParsed is true

TEST COMPLETE
